requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-multipart
alembic
mysqlclient
typing-extensions
python-dotenv
